+++ src/components/HeadingComponent.tsx
@@ -1,0 +1,9 @@
+export const HeadingComponent = () => {
+    return (
+        <div>
+            <h1 className="text-center mt-4">Heading One</h1>
+            <hr />
+        </div>
+     
+    );
+}